gke-connect-agent-generator/requirements.txt
google-cloud-gke-hub
pyyaml